requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
PyMySQL==1.1.1
sqlalchemy
groq==0.9.0
sentence-transformers==2.7.0
faiss-cpu
PyPDF2==3.0.1
numpy
pydantic>=2.9.0
pydantic-settings>=2.2.1
httpx==0.27.0
fpdf2==2.7.9
